feat: allow context-specific plugins configuration
