+++ packages/config/tests/normalize/snapshots/tests.js.md
@@ -1155,6 +1155,115 @@ Generated by [AVA](https://ava.li).
       "siteInfo": {}␊
     }`
 
+## Merges context-specific plugins
+
+> Snapshot 1
+
+    `{␊
+      "branch": "branch",␊
+      "buildDir": "packages/config/tests/normalize/fixtures/merge_netlify_toml_context",␊
+      "config": {␊
+        "build": {␊
+          "environment": {}␊
+        },␊
+        "functions": {},␊
+        "plugins": [␊
+          {␊
+            "inputs": {␊
+              "array": [␊
+                "c",␊
+                "d"␊
+              ],␊
+              "boolean": false,␊
+              "object": {␊
+                "prop": false␊
+              }␊
+            },␊
+            "origin": "config",␊
+            "package": "netlify-plugin-test"␊
+          },␊
+          {␊
+            "inputs": {␊
+              "boolean": true␊
+            },␊
+            "origin": "config",␊
+            "package": "netlify-plugin-test-two"␊
+          }␊
+        ]␊
+      },␊
+      "configPath": "packages/config/tests/normalize/fixtures/merge_netlify_toml_context/netlify.toml",␊
+      "context": "production",␊
+      "env": {␊
+        "BRANCH": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "branch"␊
+        },␊
+        "CACHED_COMMIT_REF": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "HEXADECIMAL_ID"␊
+        },␊
+        "COMMIT_REF": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "HEXADECIMAL_ID"␊
+        },␊
+        "CONTEXT": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "production"␊
+        },␊
+        "GATSBY_TELEMETRY_DISABLED": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "1"␊
+        },␊
+        "HEAD": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "branch"␊
+        },␊
+        "LANG": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "en_US.UTF-8"␊
+        },␊
+        "LANGUAGE": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "en_US:en"␊
+        },␊
+        "LC_ALL": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "en_US.UTF-8"␊
+        },␊
+        "NEXT_TELEMETRY_DISABLED": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "1"␊
+        },␊
+        "PULL_REQUEST": {␊
+          "sources": [␊
+            "general"␊
+          ],␊
+          "value": "false"␊
+        }␊
+      },␊
+      "siteInfo": {}␊
+    }`
+
 ## Merges plugins in netlify.toml and defaultConfig
 
 > Snapshot 1
